+++ ci/linux-wheels.sh
@@ -15,12 +15,13 @@ cmake .. \
   -DCMAKE_BUILD_TYPE=Release \
   -DCMAKE_INSTALL_PREFIX=/opt/smelibs \
   -DBUILD_BENCHMARKS=off \
-  -DBUILD_GUI=OFF \
+  -DBUILD_GUI=off \
   -DBUILD_CLI=off \
   -DBUILD_TESTING=off \
   -DBUILD_PYTHON_LIBRARY=off \
   -DCMAKE_PREFIX_PATH="$CMAKE_PREFIX_PATH" \
-  -DSME_EXTERNAL_SMECORE=off
+  -DSME_EXTERNAL_SMECORE=off \
+  -DSME_WITH_OPENMP=on
 make -j2 core
 make install
 cd ..